+++ scripts/mariadb-install.sh
@@ -6,30 +6,30 @@
 STACK_ROOT=$( dirname $( cd "$( dirname "$0" )" && pwd ) )
 source "${STACK_ROOT}/includes/function.inc.sh"
 
-title "MariaDB 10.1 을 설치합니다."
+title "MariaDB 10.4 을 설치합니다."
 
 
-yum_install MariaDB-server MariaDB-client MariaDB-common MariaDB-compat
+yum_install MariaDB-server MariaDB-client MariaDB-common MariaDB-compat MariaDB-shared
 
-# 메모리 선택
-if [ ${1} = "4G" ]; then
-  if [ ! -f /etc/my.cnf.d/my-innodb-heavy-4G.cnf ]; then
-    notice "MariaDB 메모리 사용량이 4GB 로 최적화되었습니다.\n설정 파일 경로) /usr/share/mysql/my-innodb-heavy-4G.cnf"
-    cp -av /usr/share/mysql/my-innodb-heavy-4G.cnf /etc/my.cnf.d/
-  fi
-fi
-if [ ${1} = "2G" ]; then
-  if [ ! -f /etc/my.cnf.d/my-huge.cnf ]; then
-    notice "MariaDB 메모리 사용량이 2GB 로 최적화되었습니다.\n설정 파일 경로) /usr/share/mysql/my-huge.cnf"
-    cp -av /usr/share/mysql/my-huge.cnf /etc/my.cnf.d/
-  fi
-fi
-if [ ${1} = "512M" ]; then
-  if [ ! -f /etc/my.cnf.d/my-large.cnf ]; then
-    notice "MariaDB 메모리 사용량이 512MB 로 최적화되었습니다.\n설정 파일 경로) /usr/share/mysql/my-large.cnf"
-    cp -av /usr/share/mysql/my-large.cnf /etc/my.cnf.d/
-  fi
-fi
+# 메모리 선택 -> MariaDB 10.3 부터 미지원
+#if [ ${1} = "4G" ]; then
+#  if [ ! -f /etc/my.cnf.d/my-innodb-heavy-4G.cnf ]; then
+#    notice "MariaDB 메모리 사용량이 4GB 로 최적화되었습니다.\n설정 파일 경로) /usr/share/mysql/my-innodb-heavy-4G.cnf"
+#    cp -av /usr/share/mysql/my-innodb-heavy-4G.cnf /etc/my.cnf.d/
+#  fi
+#fi
+#if [ ${1} = "2G" ]; then
+#  if [ ! -f /etc/my.cnf.d/my-huge.cnf ]; then
+#    notice "MariaDB 메모리 사용량이 2GB 로 최적화되었습니다.\n설정 파일 경로) /usr/share/mysql/my-huge.cnf"
+#    cp -av /usr/share/mysql/my-huge.cnf /etc/my.cnf.d/
+#  fi
+#fi
+#if [ ${1} = "512M" ]; then
+#  if [ ! -f /etc/my.cnf.d/my-large.cnf ]; then
+#    notice "MariaDB 메모리 사용량이 512MB 로 최적화되었습니다.\n설정 파일 경로) /usr/share/mysql/my-large.cnf"
+#    cp -av /usr/share/mysql/my-large.cnf /etc/my.cnf.d/
+#  fi
+#fi
 
 if [ ! -f /etc/my.cnf.d/z-php79.cnf ]; then
   notice "MariaDB 기본 캐릭터셋이 utf8mb4 로 지정되었습니다.\n설정 파일 경로) /etc/my.cnf.d/z-php79.cnf"
@@ -45,4 +45,5 @@ else
 fi
 
 # secure installation
-echo -e "\nn\n\n\n\n\n" | /usr/bin/mysql_secure_installation
+#echo -e "\nn\n\n\n\n\n" | /usr/bin/mysql_secure_installation    # <= 10.3
+echo -e "\nn\nn\n\n\n\n\n" | /usr/bin/mysql_secure_installation  # >= 10.4  https://github.com/php79/stack/issues/58